Create BuildCriteriaProvider
