+++ src/components/App.tsx
@@ -13,6 +13,7 @@ import PathwaysList from './PathwaysList';
 import CriteriaList from './CriteriaList';
 import { CurrentPathwayProvider } from './CurrentPathwayProvider';
 import { CurrentNodeProvider } from './CurrentNodeProvider';
+import { BuildCriteriaProvider } from './BuildCriteriaProvider';
 
 const App: FC = () => {
   return (
@@ -22,25 +23,27 @@ const App: FC = () => {
           <PathwayProvider>
             <CurrentNodeProvider>
               <CriteriaProvider>
-                <Router>
-                  <Switch>
-                    <Route path="/builder/:id/node/:nodeId">
-                      <BuilderRoute />
-                    </Route>
-                    <Route path="/builder/:id">
-                      <BuilderRoute />
-                    </Route>
-                    <Route path="/">
-                      <Header />
-                      <Tabs
-                        tabs={[
-                          { label: 'Pathway', component: <PathwaysList /> },
-                          { label: 'Criteria', component: <CriteriaList /> }
-                        ]}
-                      />
-                    </Route>
-                  </Switch>
-                </Router>
+                <BuildCriteriaProvider>
+                  <Router>
+                    <Switch>
+                      <Route path="/builder/:id/node/:nodeId">
+                        <BuilderRoute />
+                      </Route>
+                      <Route path="/builder/:id">
+                        <BuilderRoute />
+                      </Route>
+                      <Route path="/">
+                        <Header />
+                        <Tabs
+                          tabs={[
+                            { label: 'Pathway', component: <PathwaysList /> },
+                            { label: 'Criteria', component: <CriteriaList /> }
+                          ]}
+                        />
+                      </Route>
+                    </Switch>
+                  </Router>
+                </BuildCriteriaProvider>
               </CriteriaProvider>
             </CurrentNodeProvider>
           </PathwayProvider>
