+++ src/components/CriteriaProvider.tsx
@@ -24,10 +24,8 @@ interface Criteria {
 
 interface CriteriaContextInterface {
   criteria: Criteria[];
-  buildCriteriaNodeId: string;
   addCriteria: (file: File) => void;
   deleteCriteria: (id: string) => void;
-  updateBuildCriteriaNodeId: (id: string) => void;
 }
 
 export const CriteriaContext = createContext<CriteriaContextInterface>(
@@ -70,7 +68,6 @@ function jsonToCriteria(rawElm: string): Criteria | undefined {
 
 export const CriteriaProvider: FC<CriteriaProviderProps> = memo(({ children }) => {
   const [criteria, setCriteria] = useState<Criteria[]>([]);
-  const [buildCriteriaNodeId, setBuildCriteriaNodeId] = useState<string>('');
   const service = useGetService<Criteria>(config.get('demoCriteria'));
   const payload = (service as ServiceLoaded<Criteria[]>).payload;
 
@@ -101,21 +98,12 @@ export const CriteriaProvider: FC<CriteriaProviderProps> = memo(({ children }) =
     setCriteria(currentCriteria => currentCriteria.filter(criteria => criteria.id !== id));
   }, []);
 
-  const updateBuildCriteriaNodeId = useCallback(
-    (id: string) => {
-      setBuildCriteriaNodeId(id);
-    },
-    [setBuildCriteriaNodeId]
-  );
-
   return (
     <CriteriaContext.Provider
       value={{
         criteria,
-        buildCriteriaNodeId,
         addCriteria,
-        deleteCriteria,
-        updateBuildCriteriaNodeId
+        deleteCriteria
       }}
     >
       {children}
